% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Litbg.R
\name{BG.Lit}
\alias{BG.Lit}
\title{Literature Percent Bare Ground}
\usage{
BG.Lit(x, author)
}
\arguments{
\item{x}{From King (1998): Percent bare ground as a cover percentage measured by the Daubenmire method (Daubenmire 1959).}

\item{author}{The author of the paper from which the hsi function is estimated, one of c("simms", "guthery").}
}
\value{
Returns the relative HSI value
}
\description{
Calculates the partial HSI given the percent bare ground.
}
\author{
Dominic LaRoche
}

