% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTbg.R
\name{BG.RT}
\alias{BG.RT}
\title{Dr. Roy Tomlinson Bare Ground}
\usage{
BG.RT(x)
}
\arguments{
\item{x}{Bare Ground measured as the proportion of surface area not occupied by stems or other obstructions on a typical  home range (10.9 ha).}
}
\value{
Returns the relative HSI value
}
\description{
Calculates the partial HSI given the bare ground.
}

