% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGDCbg.R
\name{BG.SGDC}
\alias{BG.SGDC}
\title{Dr.s Sally Gall and Dan Cohan Bare Ground}
\usage{
BG.SGDC(x)
}
\arguments{
\item{x}{Bare ground measured as the average canopy cover of bare ground.  Bare ground should be in the form of a matrix interspersed with other canopy components.}
}
\value{
Returns the relative HSI value
}
\description{
Calculates the partial HSI given the bare ground.
}
\author{
Dominic LaRoche
}

