% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Litcv.R
\name{CV.Lit}
\alias{CV.Lit}
\title{Literature Cone of Vulnerability}
\usage{
CV.Lit(x, AZ = TRUE)
}
\arguments{
\item{x}{From Guthery et al.  (2000):   Visual cover measured as a cone of vulnerability (as described in Kopp (1998)) measured in millions of cubic meters around a random point.}

\item{AZ}{If TRUE, the partial HSI for Arizona is returned.  Otherwise, the partial HSO for Mexico is returned.}
}
\value{
Returns the relative HSI value
}
\description{
Calculates the partial HSI given the cone of vulnerability.
}

