% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Litdv.R
\name{DV.Lit}
\alias{DV.Lit}
\title{Literature Disc of Vulnerability}
\usage{
DV.Lit(x)
}
\arguments{
\item{x}{From Guthery et al. (2000): Visual cover measured as a disc of vulnerability (as described in Kopp (1998)) around a random point.}
}
\value{
Returns the relative HSI value
}
\description{
Calculates the partial HSI given the disc of vulnerability.
}
\author{
Dominic LaRoche
}

