% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JGfc.R
\name{FC.Goodwin}
\alias{FC.Goodwin}
\title{John Goodwin Forb Cover}
\usage{
FC.Goodwin(x, AZ = TRUE)
}
\arguments{
\item{x}{Forb cover measured as the average canopy cover of forbs.  Suitability differs in Arizona and Mexico.}

\item{AZ}{True or False.  Is the habitat in Arizona (TRUE) or Mexico (FALSE).  Defaults to Arizona.}
}
\value{
Returns the relative HSI value
}
\description{
Calculates the partial HSI given the forb cover.
}
\author{
Dominic LaRoche
}

