% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Litfc.R
\name{FC.Lit}
\alias{FC.Lit}
\title{Literature Forb Cover}
\usage{
FC.Lit(x, author)
}
\arguments{
\item{x}{Percent aerial forb cover measured as in a Daubenmire plot (Daubenmire 1959).}

\item{author}{The paper from which to calculate the partion HSI, c("king", "simms")}
}
\value{
Returns the relative HSI value
}
\description{
Calculates the partial HSI given the forb cover.
}
\author{
Dominic LaRoche
}

