% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTfc.R
\name{FC.RT}
\alias{FC.RT}
\title{Dr. Roy Tomlinson Forb Cover}
\usage{
FC.RT(x, summer = TRUE)
}
\arguments{
\item{x}{Forb Cover measured as the average total canopy cover of forbs on a on a typical home range (10.9 ha).  Suitability of forb cover differs in winter and summer.}

\item{summer}{If TRUE then returns the partial HSI for summer, otherwise it returns the partial HSI for winter (default = TRUE).}
}
\value{
Returns the relative HSI value
}
\description{
Calculates the partial HSI given the forb cover.
}
\author{
Dominic LaRoche
}

