% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGDCfc.R
\name{FC.SGDC}
\alias{FC.SGDC}
\title{Dr.s Sally Gall and Dan Cohan Forb Cover}
\usage{
FC.SGDC(x, summer = TRUE)
}
\arguments{
\item{x}{Forb cover measured as the average percent canopy cover dominated by forbs.  The optimal canopy cover of forbs differs between the fall/winter and spring/summer.}

\item{summer}{If TRUE then returns the partial HSI for summer, otherwise it returns the partial HSI for winter (default = TRUE).}
}
\value{
Returns the relative HSI value
}
\description{
Calculates the partial HSI given the forb cover.
}
\author{
Dominic LaRoche
}

