% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEfd.R
\name{FD.Ellis}
\alias{FD.Ellis}
\title{Dr. Dave Ellis Forb Diversity}
\usage{
FD.Ellis(x)
}
\arguments{
\item{x}{Forb Diversity measured as the total number of forb species on a given home range throughout the year.}
}
\value{
Returns the relative HSI value
}
\description{
Calculates the partial HSI given the forb diversity.
}
\author{
Dominic LaRoche
}

