% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JGfd.R
\name{FD.Goodwin}
\alias{FD.Goodwin}
\title{John Goodwin Forb Diversity}
\usage{
FD.Goodwin(x)
}
\arguments{
\item{x}{Forb Diversity measured as the total number of forb species found in reasonable abundance on a given home range throughout the year.}
}
\value{
Returns the relative HSI value
}
\description{
Calculates the partial HSI given the forb diversity.
}

