% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Litfd.R
\name{FD.Lit}
\alias{FD.Lit}
\title{Literature Forb Diversity}
\usage{
FD.Lit(x)
}
\arguments{
\item{x}{From Simms (1989): Forb diversity as measured by the mean number of forb species as measured by the Daubenmire method (Daubenmire 1959).}
}
\value{
Returns the relative HSI value
}
\description{
Calculates the partial HSI given the forb diversity.
}
\author{
Dominic LaRoche
}

