% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTfd.R
\name{FD.RT}
\alias{FD.RT}
\title{Dr. Roy Tomlinson Forb Diversity}
\usage{
FD.RT(x)
}
\arguments{
\item{x}{Forb Diversity measured as the total number of forb species  on a typical  home range (10.9 ha).}
}
\value{
Returns the relative HSI value
}
\description{
Calculates the partial HSI given the forb diversity.
}

