% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGDCfd.R
\name{FD.SGDC}
\alias{FD.SGDC}
\title{Dr.s Sally Gall and Dan Cohan Forb Diversity}
\usage{
FD.SGDC(x)
}
\arguments{
\item{x}{Forb Diversity measured as the total number of forb species on a typical  home range (10.9 ha) throughout the year.}
}
\value{
Returns the relative HSI value
}
\description{
Calculates the partial HSI given the forb diversity.
}

