% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEfh.R
\name{FH.Ellis}
\alias{FH.Ellis}
\title{Dr. Dave Ellis Forb Height}
\usage{
FH.Ellis(x)
}
\arguments{
\item{x}{Forb Height measured as the average height of Forbs on a given home range in centimeters.}
}
\value{
Returns the relative HSI value
}
\description{
Calculates the partial HSI given the forb height.
}

