% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTfh.R
\name{FH.RT}
\alias{FH.RT}
\title{Dr. Roy Tomlinson Forb Height}
\usage{
FH.RT(x)
}
\arguments{
\item{x}{Forb Height measured as the average height of forbs on a typical  home range (10.9 ha).}
}
\value{
Returns the relative HSI value
}
\description{
Calculates the partial HSI given the forb height.
}
\author{
Dominic LaRoche
}

