% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGDCfh.R
\name{FH.SGDC}
\alias{FH.SGDC}
\title{Dr.s Sally Gall and Dan Cohan Forb Height}
\usage{
FH.SGDC(x, summer = TRUE)
}
\arguments{
\item{x}{Forb height measured as the average height of forbs.  Optimal forb height differs between the spring/summer and the fall/winter.}

\item{summer}{If TRUE then returns the partial HSI for summer, otherwise it returns the partial HSI for winter (default = TRUE).}
}
\value{
Returns the relative HSI value
}
\description{
Calculates the partial HSI given the forb height.
}
\author{
Dominic LaRoche
}

