% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JGgc.R
\name{GC.Goodwin}
\alias{GC.Goodwin}
\title{John Goodwin Grass Cover}
\usage{
GC.Goodwin(x, AZ = TRUE)
}
\arguments{
\item{x}{Grass cover measured as the average canopy cover of grass.  Suitability differs between Arizona and Mexico.}

\item{AZ}{True or False.  Is the habitat in Arizona (TRUE) or Mexico (FALSE).  Defaults to Arizona.}
}
\value{
Returns the relative HSI value
}
\description{
Calculates the partial HSI given the grass cover.
}

