% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Litgc.R
\name{GC.Lit}
\alias{GC.Lit}
\title{Literature Grass Cover}
\usage{
GC.Lit(x)
}
\arguments{
\item{x}{From Simms (1989): Percent aerial grass cover measured as in a Daubenmire plot (Daubenmire 1959).}
}
\value{
Returns the relative HSI value
}
\description{
Calculates the partial HSI given the grass cover.
}
\author{
Dominic LaRoche
}

