% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTgc.R
\name{GC.RT}
\alias{GC.RT}
\title{Dr. Roy Tomlinson Grass Cover}
\usage{
GC.RT(x)
}
\arguments{
\item{x}{Grass Canopy Cover measured from above the grass canopy as the amount of ground covered by grass foliage  on a typical  home range (10.9 ha)}
}
\value{
Returns the relative HSI value
}
\description{
Calculates the partial HSI given the grass cover.
}
\author{
Dominic LaRoche
}

