% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGDCgc.R
\name{GC.SGDC}
\alias{GC.SGDC}
\title{Dr.s Sally Gall and Dan Cohan Grass Cover}
\usage{
GC.SGDC(x, perennial = TRUE)
}
\arguments{
\item{x}{Grass cover measured as the percent canopy cover of grass. The optimal canopy cover of grass differs between perennial and annual grasses.}

\item{perennial}{If TRUE then returns the partial HSI for perrennial grass cover, otherwise it returns the partial HSI for annual grass cover (default = TRUE).}
}
\value{
Returns the relative HSI value
}
\description{
Calculates the partial HSI given the grass cover.
}

