% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEgcb.R
\name{GCB.Ellis}
\alias{GCB.Ellis}
\title{Dr. Dave Ellis Grass Basal Area}
\usage{
GCB.Ellis(x)
}
\arguments{
\item{x}{Grass basal area measured as the average area occupied by stems of grass on a given home range.}
}
\value{
Returns the relative HSI value
}
\description{
Calculates the partial HSI given the grass basal area.
}
\author{
Dominic LaRoche
}

