% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Litgcb.R
\name{GCB.Lit}
\alias{GCB.Lit}
\title{Literature Basal Grass Cover}
\usage{
GCB.Lit(x)
}
\arguments{
\item{x}{From Simms (1989): Percent basal grass cover measured as in a Daubenmire plot (Daubenmire 1959).}
}
\value{
Returns the relative HSI value
}
\description{
Calculates the partial HSI given the basal grass cover.
}

