% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEgcc.R
\name{GCC.Ellis}
\alias{GCC.Ellis}
\title{Dr. Dave Ellis Grass Canopy Cover}
\usage{
GCC.Ellis(x)
}
\arguments{
\item{x}{Grass Canopy Cover measured from above the grass canopy as the amount of ground covered by grass foliage on a given home range.}
}
\value{
Returns the relative HSI value
}
\description{
Calculates the partial HSI given the grass canopy cover.
}

