% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEgcs.R
\name{GCS.Ellis}
\alias{GCS.Ellis}
\title{Dr. Dave Ellis Grass Side Cover}
\usage{
GCS.Ellis(x)
}
\arguments{
\item{x}{Grass Cover from the side measured as the average amount of distance (in meters) until complete visual obstruction on a given home range.}
}
\value{
Returns the relative HSI value
}
\description{
Calculates the partial HSI given the grass side cover.
}
\author{
Dominic LaRoche
}

