% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEgd.R
\name{GD.Ellis}
\alias{GD.Ellis}
\title{Dr. Dave Ellis Grass Diversity}
\usage{
GD.Ellis(x)
}
\arguments{
\item{x}{Grass Diversity measured as the total number of both annual and perennial grass species on a given home range throughout the year.}
}
\value{
Returns the relative HSI value
}
\description{
Calculates the partial HSI given the grass diversity.
}

