% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JGgd.R
\name{GD.Goodwin}
\alias{GD.Goodwin}
\title{John Goodwin Grass Diversity}
\usage{
GD.Goodwin(x, AZ = TRUE)
}
\arguments{
\item{x}{Grass Diversity measured as the total number of both annual and perennial grass species found in reasonable abundance on a given home range throughout the year.  Optimal levels of grass diversity differ in Arizona and Mexico.}

\item{AZ}{True or False.  Is the habitat in Arizona (TRUE) or Mexico (FALSE).  Defaults to Arizona.}
}
\value{
Returns the relative HSI value
}
\description{
Calculates the partial HSI given the grass diversity.
}

