% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Litgd.R
\name{GD.Lit}
\alias{GD.Lit}
\title{Literature Grass Diversity}
\usage{
GD.Lit(x)
}
\arguments{
\item{x}{From Simms (1989): Grass diversity as measured by the mean number of grass species as measured by the Daubenmire method (Daubenmire 1959).}
}
\value{
Returns the relative HSI value
}
\description{
Calculates the partial HSI given the grass diversity.
}
\author{
Dominic LaRoche
}

