% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTgd.R
\name{GD.RT}
\alias{GD.RT}
\title{Dr. Roy Tomlinson}
\usage{
GD.RT(x)
}
\arguments{
\item{x}{Grass Diversity measured as the total number of both annual and perennial grass species  on a typical  home range (10.9 ha).}
}
\value{
Returns the relative HSI value
}
\description{
Calculates the partial HSI given the grass diversity.
}
\author{
Dominic LaRoche
}

