% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGDCgd.R
\name{GD.SGDC}
\alias{GD.SGDC}
\title{Dr.s Sally Gall and Dan Cohan Grass Diversity}
\usage{
GD.SGDC(x, perennial = TRUE)
}
\arguments{
\item{x}{Grass diversity measured as the total number of grass species found on a typical home range (10.9 ha).  The optimal number of species differs between perennial and annual grasses.}

\item{perennial}{If TRUE then returns the partial HSI for perrennial grass cover, otherwise it returns the partial HSI for annual grass diversity (default = TRUE).}
}
\value{
Returns the relative HSI value
}
\description{
Calculates the partial HSI given the grass diversity.
}
\author{
Dominic LaRoche
}

