% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEgh.R
\name{GH.Ellis}
\alias{GH.Ellis}
\title{Dr. Dave Ellis Grass Height}
\usage{
GH.Ellis(x)
}
\arguments{
\item{x}{Grass Height measured as the average height of grass on a given home range (in meters).}
}
\value{
Returns the relative HSI value
}
\description{
Calculates the partial HSI given the grass height.
}
\author{
Dominic LaRoche
}

