% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTgh.R
\name{GH.RT}
\alias{GH.RT}
\title{Dr. Roy Tomlinson Grass Height}
\usage{
GH.RT(x)
}
\arguments{
\item{x}{Grass Height measured as the average height of grass  on a typical  home range (10.9 ha).}
}
\value{
Returns the relative HSI value
}
\description{
Calculates the partial HSI given the grass height.
}

