% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGDCgh.R
\name{GH.SGDC}
\alias{GH.SGDC}
\title{Dr.s Sally Gall and Dan Cohan Grass Height}
\usage{
GH.SGDC(x, expert = TRUE)
}
\arguments{
\item{x}{Grass height measured as the average height of grass on a typical home range (10.9 ha).  The two experts differed on their assessment of optimal grass height.}

\item{expert}{1 or 2. If 1 the "expert 1" model is used (see LaRoche and Conway 2013), if 2 the "expert 2" model is used}
}
\value{
Returns the relative HSI value
}
\description{
Calculates the partial HSI given the grass height.
}

