% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lithsc.R
\name{HSC.Lit}
\alias{HSC.Lit}
\title{Literature Half-Shrub Cover}
\usage{
HSC.Lit(x)
}
\arguments{
\item{x}{From Simms (1989):  Percent half-shrub cover measured by the Daubenmire method (Daubenmire 1959).}
}
\value{
Returns the relative HSI value
}
\description{
Calculates the partial HSI given the percent half-shrub cover.
}
\author{
Dominic LaRoche
}

