% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEhsi.R
\name{HSI.Ellis}
\alias{HSI.Ellis}
\title{Dave Ellis HSI Function}
\usage{
HSI.Ellis(fh, gh, sh, tc, gcc, gcs, gcb, fd, gd, sd, totc, Summer = TRUE)
}
\arguments{
\item{fh}{Forb Height measured as the average height of Forbs on a given home range in centimeters.}

\item{gh}{Grass Height measured as the average height of grass on a given home range (in meters).}

\item{sh}{Shrub Height measured as the average height of grass on a given home range (in meters).}

\item{tc}{Tree cover measured as the percent canopy cover of trees on a given home range.}

\item{gcc}{Grass Canopy Cover measured from above the grass canopy as the amount of ground covered by grass foliage on a given home range.}

\item{gcs}{Grass Cover from the side measured as the average amount of distance (in meters) until complete visual obstruction on a given home range.}

\item{gcb}{Grass basal area measured as the average area occupied by stems of grass on a given home range.}

\item{fd}{Forb Diversity measured as the total number of forb species on a given home range throughout the year.}

\item{gd}{Diversity measured as the total number of both annual and perennial grass species on a given home range throughout the year.}

\item{sd}{Shrub diversity measured as the total number of shrub species on  a given home range throughout the year.}

\item{totc}{Total Cover measured as the average total canopy cover of all vegetation (and brush piles) on a given home range.  Suitability of total cover differs in winter and summer.}

\item{Summer}{TRUE if measured in summer (default = TRUE).}
}
\value{
Returns the relative HSI value
}
\description{
Calculates the HSI given the values of forb cover, grass cover, shrub cover, tree cover, forb diversity, and grass diversity.
}

