% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JGhsi.R
\name{HSI.goodwin}
\alias{HSI.goodwin}
\title{John Goodwin HSI Function}
\usage{
HSI.goodwin(fd, fc, gd, gc, tc, sc, AZ = TRUE)
}
\arguments{
\item{fd}{The number of forb species per acre, as measured on a line intercept or frame method.}

\item{fc}{Forb cover measured as the average canopy cover of forbs.}

\item{gd}{Grass Diversity measured as the total number of both annual and perennial grass species found in reasonable abundance on a given home range throughout the year.}

\item{gc}{Grass cover measured as the average canopy cover of grass.}

\item{tc}{Tree cover measured as the average canopy cover of trees.}

\item{sc}{Shrub cover measured as the average canopy cover of shrubs throughout the year.}

\item{AZ}{True or False.  Is the habitat in Arizona (TRUE) or Mexico (FALSE).  Defaults to Arizona.}
}
\value{
Returns the relative HSI value
}
\description{
Calculates the HSI given the values of forb cover, grass cover, shrub cover, tree cover, forb diversity, and grass diversity.
}

