% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lithsi.R
\name{HSI.Lit}
\alias{HSI.Lit}
\title{Literature HSI Function}
\usage{
HSI.Lit(wsd, wc, tc, nt, vc, dv, cv, nf, fd, fc, gc, gcb, llg, hsc, bg, ot, AZ = TRUE)
}
\arguments{
\item{wsd}{From King (1998): The mean number of woody stems >1m tall per 200 square meters.}

\item{wc}{Percent woody cover as measured in a line intercept method (Canfield 1941).}

\item{tc}{From Simms (1989): Percent tree and shrub cover as measured by a line intercept method (Canfield 1941).}

\item{nt}{From Simms (1989): Number of trees with a height between 0 and 5 meters per hectare.}

\item{vc}{From Simms (1989): Visual cover measured as percent visual obstruction of a vertical range pole (Robel et al. 1970) at a height of 1m and distance of 4m.}

\item{dv}{From Guthery et al. (2000): Visual cover measured as a disc of vulnerability (as described in Kopp (1998)) around a random point.}

\item{cv}{From Guthery et al.  (2000):   Visual cover measured as a cone of vulnerability (as described in Kopp (1998)) measured in millions of cubic meters around a random point.}

\item{nf}{From King (1998): The number of forbs counted per 1000 square centimeters (as in a Daubenmire frame (Daubenmire 1959)).}

\item{fd}{From Simms (1989): Forb diversity as measured by the mean number of forb species as measured by the Daubenmire method (Daubenmire 1959).}

\item{fc}{Percent aerial forb cover measured as in a Daubenmire plot (Daubenmire 1959).}

\item{gc}{From Simms (1989): Percent aerial grass cover measured as in a Daubenmire plot (Daubenmire 1959).}

\item{gcb}{From Simms (1989): Percent basal grass cover measured as in a Daubenmire plot (Daubenmire 1959).}

\item{llg}{From Simms (1989): Percent cover of Lehmann’s lovegrass measured by the Daubenmire method (Daubenmire 1959).}

\item{hsc}{From Simms (1989):  Percent half-shrub cover measured by the Daubenmire method (Daubenmire 1959).}

\item{bg}{From King (1998): Percent bare ground as a cover percentage measured by the Daubenmire method (Daubenmire 1959).}

\item{ot}{From Guthery et al. (2001): Operative temperature measured as described in Forrester et al. (1998).}

\item{AZ}{If TRUE, the partial HSI for Arizona is returned.  Otherwise, the partial HSI for Mexico is returned.}
}
\value{
Returns the relative HSI value
}
\description{
Calculates the HSI given the values of woody stem density, woody cover, tree and shrub cover, number of trees, visual cover, disc of vulnerability, cone of vulnerability, number of forbs, forb diversity, forb cover, grass cover, basal grass cover, Lehmann's lovegrass cover, bare ground, half-shrub cover, and operative temperature.
}
\author{
Dominic LaRoche
}

