% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RThsi.R
\name{HSI.RT}
\alias{HSI.RT}
\title{Dr.s Sally Gall and Dan Cohan HSI Function}
\usage{
HSI.RT(fcs, fcw, fd, fh, gc, gd, gh, tc, bg)
}
\arguments{
\item{fcs}{Forb Cover in spring/summer measured as the average total canopy cover of forbs on a on a typical home range (10.9 ha).}

\item{fcw}{Forb cover in fall/winter measured as the average total canopy cover of forbs on a on a typical home range (10.9 ha).}

\item{fd}{Forb Diversity measured as the total number of forb species  on a typical  home range (10.9 ha).}

\item{fh}{Forb Height measured as the average height of forbs on a typical  home range (10.9 ha).}

\item{gc}{Grass Canopy Cover measured from above the grass canopy as the amount of ground covered by grass foliage  on a typical  home range (10.9 ha).}

\item{gd}{Grass Diversity measured as the total number of both annual and perennial grass species  on a typical  home range (10.9 ha).}

\item{gh}{Grass Height measured as the average height of grass  on a typical  home range (10.9 ha).}

\item{tc}{Tree and shrub cover measured as the percent canopy cover of trees on a typical  home range (10.9 ha).}

\item{bg}{Bare Ground measured as the proportion of surface area not occupied by stems or other obstructions on a typical  home range (10.9 ha).}
}
\value{
Returns the relative HSI value
}
\description{
Calculates the HSI given the values of forb cover, forb height, grass cover, grass height, shrub cover, shrub height, tree cover, bare ground, forb diversity, and grass diversity.
}

