% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGDChsi.R
\name{HSI.SGDC}
\alias{HSI.SGDC}
\title{Dr.s Sally Gall and Dan Cohan HSI Function}
\usage{
HSI.SGDC(fcs, fcw, fd, fhs, fhw, gcp, gca, gdp, gda, gh, sc, sh, tc, tca, bg, expert = 1)
}
\arguments{
\item{fcs}{Forb cover in spring/summer measured as the average percent canopy cover dominated by forbs.}

\item{fcw}{Forb cover in fall/winter measured as the average percent canopy cover dominated by forbs.}

\item{fd}{Forb Diversity measured as the total number of forb species on a typical  home range (10.9 ha) throughout the year.}

\item{fhs}{Forb height measured as the average height of forbs in spring/summer.}

\item{fhw}{Forb height measured as the average height of forbs in fall/winter.}

\item{gcp}{Perennial grass cover measured as the percent canopy cover of grass.}

\item{gca}{Annual grass cover measured as the percent canopy cover of grass.}

\item{gdp}{Perennial grass diversity measured as the total number of grass species found on a typical home range (10.9 ha).}

\item{gda}{Annual grass diversity measured as the total number of grass species found on a typical home range (10.9 ha).}

\item{gh}{Grass height measured as the average height of grass on a typical home range (10.9 ha).  The two experts differed on their assessment of optimal grass height.}

\item{sc}{Shrub cover measured as the average canopy cover of shrubs.  The two experts differed in their assessment of optimal shrub cover.}

\item{sh}{Shrub height measured as the average height of shrubs.}

\item{tc}{Upland tree cover measured as the average canopy cover of trees.}

\item{tca}{Arroyo tree cover measured as the average canopy cover of trees.}

\item{bg}{Bare ground measured as the average canopy cover of bare ground.  Bare ground should be in the form of a matrix interspersed with other canopy components.}

\item{expert}{1 or 2. If 1 the "expert 1" model is used (see LaRoche and Conway 2013), if 2 the "expert 2" model is used.}
}
\value{
Returns the relative HSI value
}
\description{
Calculates the HSI given the values of forb cover, forb height, grass cover, grass height, shrub cover, shrub height, tree cover, bare ground, forb diversity, and grass diversity.
}
\author{
Dominic LaRoche
}

