% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Litllc.R
\name{LLG.Lit}
\alias{LLG.Lit}
\title{Literature Lehmann's lovegrass}
\usage{
LLG.Lit(x)
}
\arguments{
\item{x}{From Simms (1989): Percent cover of Lehmann’s lovegrass measured by the Daubenmire method (Daubenmire 1959).}
}
\value{
Returns the relative HSI value
}
\description{
Calculates the partial HSI given the percent cover of Lehmann's lovegrass.
}

