\name{MBQhsi-package}
\alias{MBQhsi-package}
\alias{MBQhsi}
\docType{package}
\title{
\packageTitle{MBQhsi}
}
\description{
\packageDescription{MBQhsi}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{MBQhsi}
\packageIndices{MBQhsi}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{MBQhsi}

Maintainer: \packageMaintainer{MBQhsi}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
