% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MHbg.R
\name{MH.bg}
\alias{MH.bg}
\title{Mary Hunnicut Bare Ground HSI}
\usage{
MH.bg(x)
}
\arguments{
\item{x}{The proportion of bare ground (in decimal form)}
}
\value{
Returns the relative HSI value
}
\description{
Calculates the partial HSI given a proportion of bare ground
}
\details{
Amount of bare ground, including that under vegetation. Must be capable of being traversed by a bobwhite. Measured as one would on Daubenmire transect.
}

