% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MHeh.R
\name{MH.eh}
\alias{MH.eh}
\title{Mary Hunnicut Layered Vegetation HSI}
\usage{
MH.eh(x)
}
\arguments{
\item{x}{A 1 or 0 indicating the presence or absence of an ecotone.}
}
\value{
Returns the relative HSI value
}
\description{
Calculates the partial HSI given the presence of an ecotone.
}
\details{
The presence or absence of an ecotone between drainage and upland grass habitat on each acre.
}
\author{
Dominic LaRoche
}

