% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MHfcgc.R
\name{MH.fcgc}
\alias{MH.fcgc}
\title{Mary Hunnicut forb cover/ grass cover HSI}
\usage{
MH.fcgc(x)
}
\arguments{
\item{x}{Proportion of green forbs or grass measured as one would in a Daubnenmire frame (per acre).}
}
\value{
Returns the relative HSI value
}
\description{
Calculates the partial HSI given a the cover of either green forbs or grass (same function for both).
}

