% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MHfd.R
\name{MH.fd}
\alias{MH.fd}
\title{Mary Hunnicut Forb Diversity HSI}
\usage{
MH.fd(x)
}
\arguments{
\item{x}{The number of forb species per acre, as measured on a line intercept or frame method.}
}
\value{
Returns the relative HSI value
}
\description{
Calculates the partial HSI given a the forb diversity.
}

