% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MHgd.R
\name{MH.gd}
\alias{MH.gd}
\title{Mary Hunnicut Grass Diversity HSI}
\usage{
MH.gd(x)
}
\arguments{
\item{x}{The number of native grass species per acre as counted on a line intercept.}
}
\value{
Returns the relative HSI value
}
\description{
Calculates the partial HSI given a the grass diversity.
}

