% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MHgr.R
\name{MH.gr}
\alias{MH.gr}
\title{Mary Hunnicut Grasshopper Abundance HSI}
\usage{
MH.gr(x, loc)
}
\arguments{
\item{x}{A vector of grasshopper abundance during the breeding season on each acre in arbitrary units}

\item{loc}{The index of the abundance measure in the supplied vector to be assigned a suitability value}
}
\value{
Returns the relative HSI value
}
\description{
Calculates the partial HSI given the grasshopper abundance
}
\details{
Grasshopper abundance during the breeding season on each acre.  The exact relationship is unknown but more is better.  The given function will assign suitability to an area relative to other measured areas.
}
\author{
Dominic LaRoche
}

