% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MHHSI.R
\name{MH.HSI}
\alias{MH.HSI}
\title{Mary Hunnicut HSI Function}
\usage{
MH.HSI(fc, gc, sc, tc, bg, nc, gr.x, gr.loc, fd, sd, gd, lv, eh)
}
\arguments{
\item{fc}{Proportion (in decimal form) of green forbs measured as one would in a Daubnenmire frame (per acre).}

\item{gc}{Proportion (in decimal form) of green grass measured as one would in a Daubnenmire frame (per acre).}

\item{sc}{The proportion of shrub cover (in decimal form).}

\item{tc}{Canopy Cover (\%) of trees as measured by a densitometer or aerial photo over each acre of ground.}

\item{bg}{The proportion of bare ground (in decimal form).}

\item{nc}{The number of appropriate nest sites, as counted in a frame or intercept technique, per acre.}

\item{gr.x}{A vector of grasshopper abundance during the breeding season on each acre in arbitrary units}

\item{gr.loc}{The index of the grasshopper abundance measure in the supplied vector of grasshopper abundances to be assigned a suitability value.}

\item{fd}{The number of forb species per acre, as measured on a line intercept or frame method.}

\item{sd}{The number of shrub and tree species per acre as counted on a line transect or quadrant.}

\item{gd}{The number of native grass species per acre as counted on a line intercept.}

\item{lv}{A 1 or 0 indicating the presence or absence of multi-layered vegetation on each acre.}

\item{eh}{A 1 or 0 indicating the presence or absence of an ecotone.}
}
\value{
Returns the relative HSI value
}
\description{
Calculates the HSI given the values of forb cover, grass cover, shrub cover, tree cover, bare ground, nest sites, grasshopper abundance, forb diversity, shrub diversity, grass diversity, layered vegetation, and ecotone presence.
}
\author{
Dominic LaRoche
}

