% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MHlv.R
\name{MH.lv}
\alias{MH.lv}
\title{Mary Hunnicut Layered Vegetation HSI}
\usage{
MH.lv(x)
}
\arguments{
\item{x}{A 1 or 0 indicating the presence or absence of multi-layered vegetation on each acre.}
}
\value{
Returns the relative HSI value
}
\description{
Calculates the partial HSI given the presence of multi-layered vegetation.
}
\details{
The presence or absence of multi-layered vegetation on each acre.
}
\author{
Dominic LaRoche
}

