% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MHnc.R
\name{MH.nc}
\alias{MH.nc}
\title{Mary Hunnicut Nest Site HSI}
\usage{
MH.nc(x)
}
\arguments{
\item{x}{The number of appropriate nest sites, as counted in a frame or intercept technique, per acre.}
}
\value{
Returns the relative HSI value
}
\description{
Calculates the partial HSI given a the number of appropriate nest sites
}
\details{
Nests are typically located in a 9 inch (23cm) round clump of grass but can be located at the base of a shrub or under other similarly dense vegetation.  In flight pens, masked bobwhites will even use artificial structures for nesting cover.  Appropriate nest sites should be available at the rate of approximately 300-600 per acre.
}
\author{
Dominic LaRoche
}

