% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MHsc.R
\name{MH.sc}
\alias{MH.sc}
\title{Mary Hunnicut Shrub Cover HSI}
\usage{
MH.sc(x)
}
\arguments{
\item{x}{The proportion of shrub cover (in decimal form)}
}
\value{
Returns the relative HSI value
}
\description{
Calculates the partial HSI given a proportion of shrub cover
}
\details{
Cover of woody shrubs as measured by aerial photo, or line intercept method, within each acre of ground.
}
\author{
Dominic LaRoche
}

