% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MHstd.R
\name{MH.std}
\alias{MH.std}
\title{Mary Hunnicut Shrub and Tree Diversity}
\usage{
MH.std(x)
}
\arguments{
\item{x}{The number of shrub and tree species, per acre, as counted on a line transect or quadrant.}
}
\value{
Returns the relative HSI value
}
\description{
Calculates the partial HSI given the tree and shrub diversity.
}

