% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MHtc.R
\name{MH.tc}
\alias{MH.tc}
\title{Mary Hunnicut Tree Cover HSI}
\usage{
MH.tc(x)
}
\arguments{
\item{x}{Canopy Cover of trees as measured by a densitometer or aerial photo over each acre of ground.}
}
\value{
Returns the relative HSI value
}
\description{
Calculates the partial HSI given the tree and shrub diversity.
}
\author{
Dominic LaRoche
}

