% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Litnfk.R
\name{NF.Lit}
\alias{NF.Lit}
\title{Literature Number of Forbs}
\usage{
NF.Lit(x)
}
\arguments{
\item{x}{From King (1998): The number of forbs counted per 1000 square centimeters (as in a Daubenmire frame (Daubenmire 1959)).}
}
\value{
Returns the relative HSI value
}
\description{
Calculates the partial HSI given the number of forbs.
}

