% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Litnt.R
\name{NT.Lit}
\alias{NT.Lit}
\title{Literature Number of Trees}
\usage{
NT.Lit(x)
}
\arguments{
\item{x}{From Simms (1989): Number of trees with a height between 0 and 5 meters per hectare.}
}
\value{
Returns the relative HSI value
}
\description{
Calculates the partial HSI given the number of trees.
}
\author{
Dominic LaRoche
}

