% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Litot.R
\name{OT.Lit}
\alias{OT.Lit}
\title{Literature Operative Temperature}
\usage{
OT.Lit(x)
}
\arguments{
\item{x}{From Guthery et al. (2001): Operative temperature measured as described in Forrester et al. (1998).}
}
\value{
Returns the relative HSI value
}
\description{
Calculates the partial HSI given the percent bare ground.
}

