% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGDCsc.R
\name{SC.SGDC}
\alias{SC.SGDC}
\title{Dr.s Sally Gall and Dan Cohan Shrub Cover}
\usage{
SC.SGDC(x, expert = 1)
}
\arguments{
\item{x}{Shrub cover measured as the average canopy cover of shrubs.  The two experts differed in their assessment of optimal shrub cover.}

\item{expert}{1 or 2. If 1 the "expert 1" model is used (see LaRoche and Conway 2013), if 2 the "expert 2" model is used}
}
\value{
Returns the relative HSI value
}
\description{
Calculates the partial HSI given the shrub cover.
}
\author{
Dominic LaRoche
}

