% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEsd.R
\name{SD.Ellis}
\alias{SD.Ellis}
\title{Dr. Dave Ellis Shrub Diversity}
\usage{
SD.Ellis(x)
}
\arguments{
\item{x}{Shrub diversity measured as the total number of shrub species on  a given home range throughout the year.}
}
\value{
Returns the relative HSI value
}
\description{
Calculates the partial HSI given the shrub diversity.
}

