% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEsh.R
\name{SH.Ellis}
\alias{SH.Ellis}
\title{Dr. Dave Ellis Shrub Height}
\usage{
SH.Ellis(x)
}
\arguments{
\item{x}{Shrub Height measured as the average height of grass on a given home range (in meters).}
}
\value{
Returns the relative HSI value
}
\description{
Calculates the partial HSI given the shrub height.
}
\author{
Dominic LaRoche
}

