% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGDCsh.R
\name{SH.SGDC}
\alias{SH.SGDC}
\title{Dr.s Sally Gall and Dan Cohan Shrub Height}
\usage{
SH.SGDC(x)
}
\arguments{
\item{x}{Shrub height measured as the average height of shrubs.}
}
\value{
Returns the relative HSI value
}
\description{
Calculates the partial HSI given the shrub height.
}

