% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEtc.R
\name{TC.Ellis}
\alias{TC.Ellis}
\title{Dr. Dave Ellis Tree Cover}
\usage{
TC.Ellis(x)
}
\arguments{
\item{x}{Tree cover measured as the percent canopy cover of trees on a given home range.}
}
\value{
Returns the relative HSI value
}
\description{
Calculates the partial HSI given the tree cover.
}
\author{
Dominic LaRoche
}

