% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JGtc.R
\name{TC.Goodwin}
\alias{TC.Goodwin}
\title{John Goodwin Tree Cover}
\usage{
TC.Goodwin(x, AZ = TRUE)
}
\arguments{
\item{x}{Tree cover measured as the average canopy cover of trees.  Suitability of tree cover differs between Arizona and Mexico.}

\item{AZ}{True or False.  Is the habitat in Arizona (TRUE) or Mexico (FALSE).  Defaults to Arizona.}
}
\value{
Returns the relative HSI value
}
\description{
Calculates the partial HSI given the tree cover.
}

