% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Littc.R
\name{TC.Lit}
\alias{TC.Lit}
\title{Literature Tree Cover}
\usage{
TC.Lit(x)
}
\arguments{
\item{x}{From Simms (1989): Percent tree and shrub cover as measured by a line intercept method (Canfield 1941).}
}
\value{
Returns the relative HSI value
}
\description{
Calculates the partial HSI given the tree  and shrub cover.
}

