% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTtc.R
\name{TC.RT}
\alias{TC.RT}
\title{Dr. Roy Tomlinson Tree Cover}
\usage{
TC.RT(x, summer = TRUE)
}
\arguments{
\item{x}{Tree and shrub cover measured as the percent canopy cover of trees on a typical  home range (10.9 ha).  Optimal tree and shrub cover differs between summer and winter. Brush piles can be incorporated to improve suitability.}

\item{summer}{If TRUE then returns the partial HSI for summer, otherwise it returns the partial HSI for winter (default = TRUE).}
}
\value{
Returns the relative HSI value
}
\description{
Calculates the partial HSI given the tree  and shrub cover.
}

