% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGDCtc.R
\name{TC.SGDC}
\alias{TC.SGDC}
\title{Dr.s Sally Gall and Dan Cohan Tree Cover}
\usage{
TC.SGDC(x, uplands = TRUE)
}
\arguments{
\item{x}{Tree cover measured as the average canopy cover of trees.  The optimal value of tree cover differs between the uplands and arroyos.}

\item{uplands}{If TRUE then returns the partial HSI for upland tree cover, otherwise it returns the partial HSI for arroyo cover (default = TRUE).}
}
\value{
Returns the relative HSI value
}
\description{
Calculates the partial HSI given the tree cover.
}

