% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEtotc.R
\name{TotC.Ellis}
\alias{TotC.Ellis}
\title{Dr. Dave Ellis Total Cover}
\usage{
TotC.Ellis(x, Summer = TRUE)
}
\arguments{
\item{x}{Total Cover measured as the average total canopy cover of all vegetation (and brush piles) on a given home range.  Suitability of total cover differs in winter and summer.}

\item{Summer}{TRUE if measured in summer (default = TRUE).}
}
\value{
Returns the relative HSI value
}
\description{
Calculates the partial HSI given the total cover.
}

