% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Litvc.R
\name{VC.Lit}
\alias{VC.Lit}
\title{Literature Visual Cover}
\usage{
VC.Lit(x)
}
\arguments{
\item{x}{From Simms (1989): Visual cover measured as percent visual obstruction of a vertical range pole (Robel et al. 1970) at a height of 1m and distance of 4m.}
}
\value{
Returns the relative HSI value
}
\description{
Calculates the partial HSI given the visual.
}
\author{
Dominic LaRoche
}

