% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Litwc.R
\name{WC.Lit}
\alias{WC.Lit}
\title{Literature Woody Cover}
\usage{
WC.Lit(x, AZ = TRUE, author)
}
\arguments{
\item{x}{Percent woody cover as measured in a line intercept method (Canfield 1941).}

\item{AZ}{If TRUE, the partial HSI for Arizona is returned. Otherwise the partial HSI for Mexico is returned. (ignored ofr author == 'king')}

\item{author}{The author of the paper of the model to be esitmated. Choose one of either "king" or "guthery".}
}
\value{
Returns the relative HSI value
}
\description{
Calculates the partial HSI given the woody cover.
}

