% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Litwsd.R
\name{WSD.Lit}
\alias{WSD.Lit}
\title{Literature Woody Stem Density}
\usage{
WSD.Lit(x)
}
\arguments{
\item{x}{From King (1998): The mean number of woody stems >1m tall per 200 square meters.}
}
\value{
Returns the relative HSI value
}
\description{
Calculates the partial HSI given the Woody Stem Density.
}

