% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MH.bg.R
\name{MH.bg}
\alias{MH.bg}
\title{Mary Hunnicut Bare Ground HSI}
\usage{
MH.bg(x)
}
\arguments{
\item{x}{The proportion of bare ground (in decimal form)}
}
\value{
Returns the relative HSI value
}
\description{
Calculates the HSI for bare ground given a proportion of bare ground
}

